Split out SidebarPane elements to different files
