+++ browser/src/Services/Sidebar/SidebarPane/index.ts
@@ -1,0 +1,2 @@
+export * from "./SidebarPane"
+export * from "./SidebarPaneWidgets"
