+++ browser/src/Services/Sidebar/SidebarStore.ts
@@ -62,6 +62,11 @@ export class SidebarManager {
             type: "SET_FOCUSED_ID",
             focusedEntryId: id,
         })
+
+        this._store.dispatch({
+            type: "SET_ACTIVE_ID",
+            activeEntryId: id,
+        })
     }
 
     public add(icon: SidebarIcon, pane: SidebarPane): void {
