+++ browser/src/Services/Sidebar/index.tsx
@@ -1,0 +1,34 @@
+import { commandManager } from "./../../Services/CommandManager"
+import { Configuration } from "./../../Services/Configuration"
+import { editorManager } from "./../../Services/EditorManager"
+import { windowManager } from "./../../Services/WindowManager"
+import { workspace } from "./../../Services/Workspace"
+
+import { ExplorerSplit } from "./../Explorer/ExplorerSplit"
+import { SidebarContentSplit } from "./SidebarContentSplit"
+import { SidebarSplit } from "./SidebarSplit"
+import { SidebarManager } from "./SidebarStore"
+
+import { SidebarPane, ContainerWidget, ItemWidget, LabelWidget } from "./SidebarPane"
+
+let _sidebarManager: SidebarManager = null
+
+export const activate = (configuration: Configuration) => {
+
+    _sidebarManager = new SidebarManager()
+
+    if (configuration.getValue("experimental.sidebar.enabled")) {
+        const leftDock = windowManager.getDock(2)
+        leftDock.addSplit(new SidebarSplit(_sidebarManager))
+        leftDock.addSplit(new SidebarContentSplit(_sidebarManager))
+
+        _sidebarManager.add("files-o", new ExplorerSplit(configuration, workspace, commandManager, editorManager))
+
+        const searchPane = new SidebarPane("oni.sidebar.search", "Search")
+        _sidebarManager.add("search", searchPane)
+
+        searchPane.set([new ContainerWidget(), new LabelWidget(), new LabelWidget(), new ItemWidget("oni.test"), new ItemWidget("oni.test2")])
+    }
+}
+
+export const getInstance = (): SidebarManager => _sidebarManager
